refactor(be): attendanceController 토론 관련 api -> DiscussionController로 분리
