+++ geulnamu-backend/src/docs/asciidoc/attendance-api-guide.adoc
@@ -82,70 +82,6 @@ include::{snippets}/attendance/view/list/response-fields.adoc[]
 
 ==== Error case
 
-=== 모임 토론 참여 희망 명단 조회
-접근 가능 권한 레벨: STAFF
-
-==== request
-include::{snippets}/attendance/discussion/http-request.adoc[]
-
-==== request-headers
-include::{snippets}/attendance/discussion/request-headers.adoc[]
-
-==== query-parameters
-include::{snippets}/attendance/discussion/query-parameters.adoc[]
-
-==== response
-include::{snippets}/attendance/discussion/http-response.adoc[]
-
-==== response-fields
-include::{snippets}/attendance/discussion/response-fields.adoc[]
-
-==== Error case
-
----
-
-=== 본인 토론 그룹 명단 조회
-접근 가능 권한 레벨: MEMBER
-
-==== request
-include::{snippets}/attendance/discussion/view/my-group/http-request.adoc[]
-
-==== path-parameters
-include::{snippets}/attendance/discussion/view/my-group/path-parameters.adoc[]
-
-==== request-headers
-include::{snippets}/attendance/discussion/view/my-group/request-headers.adoc[]
-
-==== response
-include::{snippets}/attendance/discussion/view/my-group/http-response.adoc[]
-
-==== response-fields
-include::{snippets}/attendance/discussion/view/my-group/response-fields.adoc[]
-
-==== Error case
-
----
-
-=== 전체 토론 그룹 명단 조회
-접근 가능 권한 레벨: STAFF
-
-==== request
-include::{snippets}/attendance/discussion/view/all-group/http-request.adoc[]
-
-==== request-headers
-include::{snippets}/attendance/discussion/view/all-group/request-headers.adoc[]
-
-==== query-parameters
-include::{snippets}/attendance/discussion/view/all-group/query-parameters.adoc[]
-
-==== response
-include::{snippets}/attendance/discussion/view/all-group/http-response.adoc[]
-
-==== response-fields
-include::{snippets}/attendance/discussion/view/all-group/response-fields.adoc[]
-
-==== Error case
-
 ---
 
 
@@ -219,55 +155,6 @@ include::{snippets}/attendance/modify/want-discussion/response-fields.adoc[]
 ==== Error case
 
 ---
-
-=== 토론 그룹 구성 - 수동
-접근 가능 권한 레벨: STAFF / 로그 생성
-
-==== request
-include::{snippets}/attendance/group/assign/http-request.adoc[]
-
-==== request-headers
-include::{snippets}/attendance/group/assign/request-headers.adoc[]
-
-==== query-parameters
-include::{snippets}/attendance/group/assign/query-parameters.adoc[]
-
-==== request-fields
-include::{snippets}/attendance/group/assign/request-fields.adoc[]
-
-==== response
-include::{snippets}/attendance/group/assign/http-response.adoc[]
-
-==== response-fields
-include::{snippets}/attendance/group/assign/response-fields.adoc[]
-
-==== Error case
-
----
-
-
-=== 토론 그룹 할당 - 개인
-접근 가능 권한 레벨: ADMIN / 로그 생성
-
-==== request
-include::{snippets}/attendance/group/assign/solo/http-request.adoc[]
-
-==== request-headers
-include::{snippets}/attendance/group/assign/solo/request-headers.adoc[]
-
-==== query-parameters
-include::{snippets}/attendance/group/assign/solo/query-parameters.adoc[]
-
-==== response
-include::{snippets}/attendance/group/assign/solo/http-response.adoc[]
-
-==== response-fields
-include::{snippets}/attendance/group/assign/solo/response-fields.adoc[]
-
-==== Error case
-
----
-
 
 == DELETE
