+++ geulnamu-backend/src/main/java/com/geulnamu/controller/attendance/AttendanceController.java
@@ -1,11 +1,8 @@
 package com.geulnamu.controller.attendance;
 
 import com.geulnamu.controller.attendance.dto.request.AttendanceNoteRequest;
-import com.geulnamu.controller.attendance.dto.request.AssignDiscussionGroupsRequest;
 import com.geulnamu.controller.attendance.dto.response.AttendanceInfoResponse;
-import com.geulnamu.controller.attendance.dto.response.DiscussionGroupResponse;
 import com.geulnamu.controller.attendance.dto.response.MeetingAttendanceDetailsResponse;
-import com.geulnamu.controller.shared.dto.response.MemberIdAndNameResponse;
 import com.geulnamu.domain.shared.enums.ActionType;
 import com.geulnamu.domain.shared.enums.Level;
 import com.geulnamu.infrastructure.annotation.AccessLevel;
@@ -15,11 +12,8 @@ import com.geulnamu.infrastructure.response.BaseResponse;
 import com.geulnamu.service.attendance.AttendanceService;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
@@ -52,28 +46,6 @@ public class AttendanceController {
         return BaseResponse.ofSuccess(meetingAttendanceStatusResponseList);
     }
 
-    @AccessLevel(Level.STAFF)
-    @GetMapping(value = "/discussion", name = "모임 토론 참여 희망 명단 조회")
-    public BaseResponse<List<MemberIdAndNameResponse>> getWantDiscussionMemberList(@RequestParam @Min(value = 1) Long meetingId) {
-        List<MemberIdAndNameResponse> memberIdAndNameResponsesList = attendanceService.getWantDiscussionMemberList(meetingId);
-        return BaseResponse.ofSuccess(memberIdAndNameResponsesList);
-    }
-
-    @AccessLevel(Level.MEMBER)
-    @GetMapping(value = "/discussion/{attendanceId}", name = "본인 토론 그룹 명단 조회")
-    public BaseResponse<List<MemberIdAndNameResponse>> getMyDiscussionGroupMemberList(@PathVariable @Min(value = 1) Long attendanceId,
-                                                                                      @AuthMemberId Long memberId) {
-        List<MemberIdAndNameResponse> memberIdAndNameResponsesList = attendanceService.getMyDiscussionMemberList(attendanceId, memberId);
-        return BaseResponse.ofSuccess(memberIdAndNameResponsesList);
-    }
-
-    @AccessLevel(Level.STAFF)
-    @GetMapping(value = "/discussion/all", name = "전체 토론 그룹 명단 조회")
-    public BaseResponse<List<DiscussionGroupResponse>> getAllDiscussionGroupMemberList(@NotNull(message = "모임 고유번호 필수 입력") @Min(value = 1) Long meetingId) {
-        List<DiscussionGroupResponse> discussionGroupResponseList = attendanceService.getAllDiscussionGroupMemberList(meetingId);
-        return BaseResponse.ofSuccess(discussionGroupResponseList);
-    }
-
     @AccessLevel(Level.MEMBER)
     @PatchMapping(value = "/{attendanceId}/note", name = "비고 작성")
     public BaseResponse<Void> writeNote(@PathVariable @Min(value = 1) Long attendanceId, @AuthMemberId Long memberId,
@@ -93,25 +65,6 @@ public class AttendanceController {
     @PatchMapping(value = "/{attendanceId}/want-discussion", name = "토론할래요")
     public BaseResponse<Void> wantDiscussion(@PathVariable @Min(value = 1) Long attendanceId, @AuthMemberId Long memberId) {
         attendanceService.wantDiscussion(attendanceId, memberId);
-        return BaseResponse.ofSuccess();
-    }
-
-    @LogAction(value = ActionType.DISCUSSION_GROUP_ORGANIZE, actionDomain = "attendance")
-    @AccessLevel(Level.STAFF)
-    @PatchMapping(value = "/group", name = "토론 그룹 구성 - 수동")
-    public BaseResponse<Void> manuallyAssignDiscussionGroups(@RequestParam @Min(value = 1) Long meetingId,
-                                                             @Valid @RequestBody AssignDiscussionGroupsRequest request) {
-        attendanceService.manuallyAssignDiscussionGroups(meetingId, request);
-        return BaseResponse.ofSuccess();
-    }
-
-    @LogAction(value = ActionType.DISCUSSION_GROUP_ORGANIZE_SOLO, actionDomain = "attendance")
-    @AccessLevel(Level.ADMIN)
-    @PatchMapping(value = "/group/solo", name = "토론 그룹 할당 - 개인")
-    public BaseResponse<Void> manuallyAssignDiscussionGroup(@RequestParam @Min(value = 1) Long meetingId,
-                                                            @RequestParam @Min(value = 1) Long memberId,
-                                                            @RequestParam @Min(value = 1) Integer groupNumber) {
-        attendanceService.assignMemberToDiscussionGroup(meetingId, memberId, groupNumber-1);
         return BaseResponse.ofSuccess();
     }
